+++ src/source/traits.rs
@@ -22,8 +22,6 @@ pub trait PackageRef: Debug {
     fn dependencies(&self) -> &BTreeMap<String, (DependencySpecifiers, DependencyType)>;
     /// Whether to use the new structure (`packages` folders inside the package's content folder) or the old structure (Wally-style, with linker files in the parent of the folder containing the package's contents)
     fn use_new_structure(&self) -> bool;
-    /// The target of this package
-    fn target_kind(&self) -> TargetKind;
     /// The source of this package
     fn source(&self) -> PackageSources;
 }